refactor(types): move DiceDefinition to remove circular import
