+++ src/types/dice.ts
@@ -1,0 +1,4 @@
+
+export type DiceDefinition = `${number}d${number}` |
+  `${number}d${number}+${number}` |
+  `${number}d${number}-${number}`;
